# gravitee-apim-rest-api/gravitee-apim-rest-api-portal/gravitee-apim-rest-api-portal-rest/src/main/resources/portal-openapi.yaml
openapi: 3.0.0
info:
    title: Gravitee.io Portal Rest API
    description: API dedicated to the devportal part of Gravitee
    termsOfService: terms
    contact:
        name: The Gravitee Team
        url: https://gravitee.io
        email: contact@graviteesource.com
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0
    version: "4.11.0-SNAPSHOT"
servers:
    - url: /portal/environments/{envId}
      description: The portal API for a given environment
      variables:
          envId:
              default: DEFAULT
              description: ID of the environment
...
security:
    - BasicAuth: []
    - CookieAuth: []
paths:
...
    /groups:
        get:
            tags:
                - Group
            parameters:
                - $ref: "#/components/parameters/pageNumberParam"
                - $ref: "#/components/parameters/pageSizeParam"
            summary: List groups
            description: |
                List all groups

                User must have the MANAGEMENT_GROUP[READ] permission.
            operationId: getGroups
            responses:
                200:
                    description: List of groups
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GroupsResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
    /user:
        get:
            tags:
                - User
            summary: Get the authenticated user
            description: |
                Get information about the authenticated user.
            operationId: getCurrentUser
            responses:
                200:
                    description: The current user
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
                500:
                    $ref: "#/components/responses/InternalServerError"
        put:
            tags:
                - User
            requestBody:
                description: Use to update a user.
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UserInput"
            summary: Modify current user information.
            description: |
                Modify current user information.

                Only the current user can modify his/her information.
                The user email is not modifiable.
            operationId: updateCurrentUser
            responses:
                200:
                    description: Updated user
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
                400:
                    description: Input must not be null.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                403:
                    description: Not allowed to modify another user.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
    /configuration:
        get:
            tags:
                - Portal
            summary: Get portal configuration.
            operationId: getPortalConfiguration
            description: |
                Get all the portal configuration from the platform settings.
            security:
                - {}
                - BasicAuth: []
                - CookieAuth: []
            responses:
                200:
                    description: Portal configuration
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ConfigurationResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
    /configuration/identities:
        get:
            tags:
                - Portal
            summary: Get the identity provider list.
            operationId: getPortalIdentityProviders
            description: |
                Get all the identity providers that can used in the portal.
            security:
                - {}
                - BasicAuth: []
                - CookieAuth: []
            responses:
                200:
                    description: List of identity providers
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ConfigurationIdentitiesResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
components:
...
    parameters:
...
        pageNumberParam:
            name: page
            in: query
            required: false
            description: The page number for pagination.
            schema:
                type: integer
                default: 1
        pageSizeParam:
            name: size
            in: query
            required: false
            description: |
                The number of items per page for pagination.
                If the size is 0, the response contains only metadata and returns the values as for a non-paged resource.
                If the size is -1, the response contains all datas.
            schema:
                type: integer
                default: 10

        #Apis filtering
...
    schemas:
...
        ErrorResponse:
            properties:
                errors:
                    description: List of errors
                    type: array
                    items:
                        $ref: "#/components/schemas/Error"
...
        GroupsResponse:
            properties:
                data:
                    description: List of groups.
                    type: array
                    items:
                        $ref: "#/components/schemas/Group"
                metadata:
                    $ref: "#/components/schemas/MetadataMap"
                links:
                    $ref: "#/components/schemas/Links"
...
        ConfigurationIdentitiesResponse:
            properties:
                data:
                    description: List of identity providers
                    type: array
                    items:
                        $ref: "#/components/schemas/IdentityProvider"
                metadata:
                    $ref: "#/components/schemas/MetadataMap"
                links:
                    $ref: "#/components/schemas/Links"
...
        Error:
            properties:
                status:
                    description: Status HTTP of the error
                    type: string
                message:
                    description: Message of the error
                    type: string
                code:
                    description: Technical code of the error
                    type: string
                parameters:
                    description: Parameters
                    type: object
                    additionalProperties:
                        type: string
...
        Group:
            required:
                - id
                - name
            properties:
                id:
                    description: Unique identifier of a group.
                    type: string
                name:
                    description: Name of the group.
                    type: string
...
        User:
            properties:
                id:
                    description: Unique identifier of a user.
                    type: string
                reference:
                    description: Unique reference if user comes from external source. Use for search only.
                    type: string
                first_name:
                    type: string
                last_name:
                    type: string
                display_name:
                    type: string
                email:
                    type: string
                editable_profile:
                    description: True if the user can edit the MyAccount information
                    type: boolean
                    default: false
                permissions:
                    description: Map of Permissions.
                    type: object
                    properties:
                        APPLICATION:
                            type: array
                            items:
                                type: string
                        USER:
                            type: array
                            items:
                                type: string
                customFields:
                    type: object
                    description: Values for CustomUserFields
                    additionalProperties: true
                config:
                    $ref: "#/components/schemas/UserConfig"
                _links:
                    $ref: "#/components/schemas/UserLinks"
        UserConfig:
            properties:
                management_url:
                    description: The URL of the Gravitee management UI
                    type: string
...
        IdentityProvider:
            properties:
                id:
                    description: Unique identifier of an identity provider.
                    type: string
                name:
                    description: Name of the identity provider.
                    type: string
                description:
                    description: Description of the identity provider.
                    type: string
                client_id:
                    description: ClientId of the identity provider.
                    type: string
                email_required:
                    description: true, if an email is required for this identity provider.
                    type: boolean
                type:
                    $ref: "#/components/schemas/IdentityProviderType"
                authorizationEndpoint:
                    description: Authorization endpoint of the provider.
                    type: string
                tokenIntrospectionEndpoint:
                    description: Token introspection endpoint of the provider. (Gravitee.io AM and OpenId Connect only)
                    type: string
                userLogoutEndpoint:
                    description: User logout endpoint of the provider. (Gravitee.io AM and OpenId Connect only)
                    type: string
                color:
                    description: color to display for this provider. (Gravitee.io AM and OpenId Connect only)
                    type: string
                display:
                    description: Display style of the provider. (Google only)
                    type: string
                requiredUrlParams:
                    description: Required URL params of the provider. (Google only)
                    type: array
                    items:
                        type: string
                optionalUrlParams:
                    description: Optionnal URL params of the provider. (Github and Google only)
                    type: array
                    items:
                        type: string
                scopes:
                    description: Scope list of the provider.
                    type: array
                    items:
                        type: string
...
        ConfigurationResponse:
            properties:
                portal:
                    $ref: "#/components/schemas/ConfigurationPortal"
                portalNext:
                    $ref: "#/components/schemas/ConfigurationPortalNext"
                authentication:
                    $ref: "#/components/schemas/ConfigurationAuthentication"
                scheduler:
                    $ref: "#/components/schemas/ConfigurationScheduler"
                documentation:
                    $ref: "#/components/schemas/ConfigurationDocumentation"
                plan:
                    $ref: "#/components/schemas/ConfigurationPlan"
                apiReview:
                    $ref: "#/components/schemas/Enabled"
                analytics:
                    $ref: "#/components/schemas/ConfigurationAnalytics"
                application:
                    $ref: "#/components/schemas/ConfigurationApplication"
                recaptcha:
                    $ref: "#/components/schemas/ConfigurationReCaptcha"
                alert:
                    $ref: "#/components/schemas/Enabled"
        ConfigurationPortal:
            properties:
                title:
                    description: The portal Title
                    type: string
                entrypoint:
                    description: Default entrypoint of the gateway.
                    type: string
                apikeyHeader:
                    description: Api-key Header. Used by portal to display the CURL command.
                    type: string
                kafkaSaslMechanisms:
                    description: Permitted Kafka Sasl Mechanisms. Used by portal to display allowed Kafka Sasl Mechanisms.
                    type: array
                    items:
                        type: string
                support:
                    $ref: "#/components/schemas/Enabled"
                applicationCreation:
                    $ref: "#/components/schemas/Enabled"
                userCreation:
                    $ref: "#/components/schemas/Enabled"
                apis:
                    $ref: "#/components/schemas/ConfigurationPortalApis"
                analytics:
                    $ref: "#/components/schemas/ConfigurationPortalAnalytics"
                rating:
                    $ref: "#/components/schemas/ConfigurationPortalRating"
                uploadMedia:
                    $ref: "#/components/schemas/ConfigurationPortalMedia"
                homepageTitle:
                    description: Main phrase to display on the homepage.
                    type: string
        ConfigurationPortalNext:
            properties:
                siteTitle:
                    description: Title to display in the navigation menu
                    type: string
                banner:
                    $ref: "#/components/schemas/ConfigurationPortalNextBanner"

                access:
                    $ref: "#/components/schemas/Enabled"
                catalog:
                    $ref: "#/components/schemas/ConfigurationPortalNextCatalog"
        ConfigurationPortalNextBanner:
            type: object
            description: Configuration of the banner for Portal Next
            properties:
                enabled:
                    description: Portal next is enabled
                    type: boolean
                title:
                    description: Title to display on the homepage banner.
                    type: string
                subtitle:
                    description: Subtitle to display on the homepage banner.
                    type: string
                primaryButton:
                    $ref: "#/components/schemas/BannerButton"
                secondaryButton:
                    $ref: "#/components/schemas/BannerButton"
        BannerButton:
            type: object
            description: Button displayed in the Portal Next banner
            properties:
                enabled:
                    description: Button is displayed
                    type: boolean
                label:
                    description: Button label
                    type: string
                type:
                    description: Type of link
                    enum:
                        - EXTERNAL
                target:
                    description: Target of the link
                    type: string
                visibility:
                    description: Visibility of the button
                    enum:
                        - PUBLIC
                        - PRIVATE
        ConfigurationPortalNextCatalog:
            type: object
            description: Configuration of the catalog for Portal Next
            properties:
                viewMode:
                    description: View mode used by the Catalog. If the value is invalid, the TABS view will be shown by default.
                    type: string
                    example: TABS
        ConfigurationPortalApis:
            properties:
                tilesMode:
                    $ref: "#/components/schemas/Enabled"
                documentationOnlyMode:
                    $ref: "#/components/schemas/Enabled"
                categoryMode:
                    $ref: "#/components/schemas/Enabled"
                promotedApiMode:
                    $ref: "#/components/schemas/Enabled"
                apiHeaderShowTags:
                    $ref: "#/components/schemas/Enabled"
                apiHeaderShowCategories:
                    $ref: "#/components/schemas/Enabled"
        ConfigurationPortalAnalytics:
            properties:
                enabled:
                    description: true if Google analytics is activated
                    type: boolean
                trackingId:
                    description: Tracking ID for Google analytics.
                    type: string
        ConfigurationPortalRating:
            properties:
                enabled:
                    description: true if rating service is activated
                    type: boolean
                comment:
                    $ref: "#/components/schemas/ConfigurationPortalRatingComment"
        ConfigurationPortalRatingComment:
            properties:
                mandatory:
                    description: true if a comment is required to post a rating.
                    type: boolean
        ConfigurationPortalMedia:
            properties:
                enabled:
                    description: true if uploading images is allowed
                    type: boolean
                maxSizeInBytes:
                    description: Maximum size of uploaded images (in bytes)
                    type: integer
        ConfigurationAuthentication:
            properties:
                forceLogin:
                    $ref: "#/components/schemas/Enabled"
                localLogin:
                    $ref: "#/components/schemas/Enabled"
        ConfigurationScheduler:
            properties:
                notificationsInSeconds:
                    description: Number of seconds for notification scheduler.
                    type: integer
        ConfigurationDocumentation:
            properties:
                url:
                    description: URL of the main documentation.
                    type: string
                pageNotFoundMessage:
                    description: Message to display when a page is not found.
                    type: string
        ConfigurationPlan:
            properties:
                security:
                    $ref: "#/components/schemas/ConfigurationPlanSecurity"
        ConfigurationPlanSecurity:
            properties:
                apikey:
                    $ref: "#/components/schemas/Enabled"
                sharedApiKey:
                    $ref: "#/components/schemas/Enabled"
                oauth2:
                    $ref: "#/components/schemas/Enabled"
                keyless:
                    $ref: "#/components/schemas/Enabled"
                jwt:
                    $ref: "#/components/schemas/Enabled"
                mtls:
                    $ref: "#/components/schemas/Enabled"
        ConfigurationAnalytics:
            properties:
                clientTimeout:
                    description: HTTP Client Timeout
                    type: integer
                    format: int64
        ConfigurationApplication:
            properties:
                registration:
                    $ref: "#/components/schemas/Enabled"
                types:
                    $ref: "#/components/schemas/ConfigurationApplicationTypes"
        ConfigurationApplicationTypes:
            properties:
                simple:
                    $ref: "#/components/schemas/Enabled"
                browser:
                    $ref: "#/components/schemas/Enabled"
                web:
                    $ref: "#/components/schemas/Enabled"
                native:
                    $ref: "#/components/schemas/Enabled"
                backend_to_backend:
                    $ref: "#/components/schemas/Enabled"
        ConfigurationReCaptcha:
            properties:
                enabled:
                    description: flag indication if recaptcha is enabled or not
                    type: boolean
                siteKey:
                    description: reCaptcha site key
                    type: string
...
        UserInput:
            required:
                - id
            properties:
                id:
                    description: Unique identifier of a user.
                    type: string
                avatar:
                    description: Avatar of the user.
                    type: string
                    writeOnly: true
                first_name:
                    description: Firstname of the user.
                    type: string
                    writeOnly: true
                last_name:
                    description: Lastname of the user.
                    type: string
                    writeOnly: true
                customFields:
                    type: object
                    description: Values for CustomUserFields
                    additionalProperties: true
...
        MetadataMap:
            description: Map of Map of Object
            type: object
            additionalProperties:
                type: object
                additionalProperties:
                    type: object
...
        Enabled:
            properties:
                enabled:
                    description: true, if the property is enabled
                    type: boolean
        IdentityProviderType:
            description: type of Identity Provider
            type: string
            enum:
                - GOOGLE
                - GITHUB
                - GRAVITEEIO_AM
                - OIDC
...
        Links:
            properties:
                self:
                    type: string
                first:
                    type: string
                last:
                    type: string
                prev:
                    type: string
                next:
                    type: string
...
        UserLinks:
            properties:
                self:
                    type: string
                avatar:
                    type: string
                notifications:
                    type: string
...
    responses:
        InternalServerError:
            description: Internal Server Error
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
...
    securitySchemes:
        BasicAuth:
            type: http
            scheme: basic
        CookieAuth:
            type: apiKey
            in: cookie
            name: Auth-Graviteeio-APIM